refactor small API group packages
