+++ pkg/master/storage_certificates.go
@@ -21,7 +21,7 @@ import (
 	"k8s.io/kubernetes/pkg/apis/certificates"
 	certificatesapiv1alpha1 "k8s.io/kubernetes/pkg/apis/certificates/v1alpha1"
 	"k8s.io/kubernetes/pkg/genericapiserver"
-	certificateetcd "k8s.io/kubernetes/pkg/registry/certificates/etcd"
+	certificateetcd "k8s.io/kubernetes/pkg/registry/certificates/certificates/etcd"
 )
 
 type CertificatesRESTStorageProvider struct{}
