+++ pkg/registry/apps/petset/doc.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2016 The Kubernetes Authors.
+Copyright 2015 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,6 +14,4 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package scheduledjob provides Registry interface and it's RESTStorage
-// implementation for storing ScheduledJob api objects.
-package scheduledjob // import "k8s.io/kubernetes/pkg/registry/scheduledjob"
+package petset // import "k8s.io/kubernetes/pkg/registry/apps/petset"
