+++ pkg/registry/apps/petset/etcd/etcd.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2014 The Kubernetes Authors.
+Copyright 2015 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -19,80 +19,82 @@ package etcd
 import (
 	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/api/rest"
-	"k8s.io/kubernetes/pkg/apis/autoscaling"
+	appsapi "k8s.io/kubernetes/pkg/apis/apps"
+	"k8s.io/kubernetes/pkg/registry/apps/petset"
 	"k8s.io/kubernetes/pkg/registry/cachesize"
 	"k8s.io/kubernetes/pkg/registry/generic"
 	"k8s.io/kubernetes/pkg/registry/generic/registry"
-	"k8s.io/kubernetes/pkg/registry/horizontalpodautoscaler"
 	"k8s.io/kubernetes/pkg/runtime"
 	"k8s.io/kubernetes/pkg/storage"
 )
 
+// rest implements a RESTStorage for replication controllers against etcd
 type REST struct {
 	*registry.Store
 }
 
-// NewREST returns a RESTStorage object that will work against horizontal pod autoscalers.
+// NewREST returns a RESTStorage object that will work against replication controllers.
 func NewREST(opts generic.RESTOptions) (*REST, *StatusREST) {
 	prefix := "/" + opts.ResourcePrefix
 
-	newListFunc := func() runtime.Object { return &autoscaling.HorizontalPodAutoscalerList{} }
+	newListFunc := func() runtime.Object { return &appsapi.PetSetList{} }
 	storageInterface, dFunc := opts.Decorator(
 		opts.StorageConfig,
-		cachesize.GetWatchCacheSizeByResource(cachesize.HorizontalPodAutoscalers),
-		&autoscaling.HorizontalPodAutoscaler{},
+		cachesize.GetWatchCacheSizeByResource(cachesize.PetSet),
+		&appsapi.PetSet{},
 		prefix,
-		horizontalpodautoscaler.Strategy,
+		petset.Strategy,
 		newListFunc,
 		storage.NoTriggerPublisher,
 	)
 
 	store := &registry.Store{
-		NewFunc: func() runtime.Object { return &autoscaling.HorizontalPodAutoscaler{} },
+		NewFunc: func() runtime.Object { return &appsapi.PetSet{} },
+
 		// NewListFunc returns an object capable of storing results of an etcd list.
 		NewListFunc: newListFunc,
-		// Produces a path that etcd understands, to the root of the resource
+		// Produces a petSet that etcd understands, to the root of the resource
 		// by combining the namespace in the context with the given prefix
 		KeyRootFunc: func(ctx api.Context) string {
 			return registry.NamespaceKeyRootFunc(ctx, prefix)
 		},
-		// Produces a path that etcd understands, to the resource by combining
+		// Produces a petSet that etcd understands, to the resource by combining
 		// the namespace in the context with the given prefix
 		KeyFunc: func(ctx api.Context, name string) (string, error) {
 			return registry.NamespaceKeyFunc(ctx, prefix, name)
 		},
-		// Retrieve the name field of an autoscaler
+		// Retrieve the name field of a replication controller
 		ObjectNameFunc: func(obj runtime.Object) (string, error) {
-			return obj.(*autoscaling.HorizontalPodAutoscaler).Name, nil
+			return obj.(*appsapi.PetSet).Name, nil
 		},
-		// Used to match objects based on labels/fields for list
-		PredicateFunc:           horizontalpodautoscaler.MatchAutoscaler,
-		QualifiedResource:       autoscaling.Resource("horizontalpodautoscalers"),
+		// Used to match objects based on labels/fields for list and watch
+		PredicateFunc:           petset.MatchPetSet,
+		QualifiedResource:       appsapi.Resource("petsets"),
 		EnableGarbageCollection: opts.EnableGarbageCollection,
 		DeleteCollectionWorkers: opts.DeleteCollectionWorkers,
 
-		// Used to validate autoscaler creation
-		CreateStrategy: horizontalpodautoscaler.Strategy,
+		// Used to validate controller creation
+		CreateStrategy: petset.Strategy,
 
-		// Used to validate autoscaler updates
-		UpdateStrategy: horizontalpodautoscaler.Strategy,
-		DeleteStrategy: horizontalpodautoscaler.Strategy,
+		// Used to validate controller updates
+		UpdateStrategy: petset.Strategy,
+		DeleteStrategy: petset.Strategy,
 
 		Storage:     storageInterface,
 		DestroyFunc: dFunc,
 	}
 	statusStore := *store
-	statusStore.UpdateStrategy = horizontalpodautoscaler.StatusStrategy
+	statusStore.UpdateStrategy = petset.StatusStrategy
 	return &REST{store}, &StatusREST{store: &statusStore}
 }
 
-// StatusREST implements the REST endpoint for changing the status of a daemonset
+// StatusREST implements the REST endpoint for changing the status of an petSet
 type StatusREST struct {
 	store *registry.Store
 }
 
 func (r *StatusREST) New() runtime.Object {
-	return &autoscaling.HorizontalPodAutoscaler{}
+	return &appsapi.PetSet{}
 }
 
 // Get retrieves the object from the storage. It is required to support Patch.
