+++ pkg/registry/autoscaling/horizontalpodautoscaler/etcd/etcd_test.go
