+++ pkg/registry/autoscaling/horizontalpodautoscaler/strategy_test.go
