+++ pkg/registry/batch/job/doc.go
@@ -16,4 +16,4 @@ limitations under the License.
 
 // Package job provides Registry interface and it's RESTStorage
 // implementation for storing Job api objects.
-package job // import "k8s.io/kubernetes/pkg/registry/job"
+package job // import "k8s.io/kubernetes/pkg/registry/batch/job"
