+++ pkg/registry/batch/job/strategy.go
