+++ pkg/registry/batch/job/strategy_test.go
