+++ pkg/registry/batch/scheduledjob/doc.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2015 The Kubernetes Authors.
+Copyright 2016 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,4 +14,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package petset // import "k8s.io/kubernetes/pkg/registry/petset"
+// Package scheduledjob provides Registry interface and it's RESTStorage
+// implementation for storing ScheduledJob api objects.
+package scheduledjob // import "k8s.io/kubernetes/pkg/registry/batch/scheduledjob"
