+++ pkg/registry/batch/scheduledjob/strategy.go
