+++ pkg/registry/batch/scheduledjob/strategy_test.go
