+++ pkg/registry/policy/poddisruptionbudget/doc.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2014 The Kubernetes Authors.
+Copyright 2015 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,4 +14,4 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package horizontalpodautoscaler // import "k8s.io/kubernetes/pkg/registry/horizontalpodautoscaler"
+package poddisruptionbudget // import "k8s.io/kubernetes/pkg/registry/policy/poddisruptionbudget"
