+++ pkg/registry/policy/poddisruptionbudget/etcd/etcd.go
@@ -19,82 +19,82 @@ package etcd
 import (
 	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/api/rest"
-	appsapi "k8s.io/kubernetes/pkg/apis/apps"
+	policyapi "k8s.io/kubernetes/pkg/apis/policy"
 	"k8s.io/kubernetes/pkg/registry/cachesize"
 	"k8s.io/kubernetes/pkg/registry/generic"
 	"k8s.io/kubernetes/pkg/registry/generic/registry"
-	"k8s.io/kubernetes/pkg/registry/petset"
+	"k8s.io/kubernetes/pkg/registry/policy/poddisruptionbudget"
 	"k8s.io/kubernetes/pkg/runtime"
 	"k8s.io/kubernetes/pkg/storage"
 )
 
-// rest implements a RESTStorage for replication controllers against etcd
+// rest implements a RESTStorage for pod disruption budgets against etcd
 type REST struct {
 	*registry.Store
 }
 
-// NewREST returns a RESTStorage object that will work against replication controllers.
+// NewREST returns a RESTStorage object that will work against pod disruption budgets.
 func NewREST(opts generic.RESTOptions) (*REST, *StatusREST) {
 	prefix := "/" + opts.ResourcePrefix
 
-	newListFunc := func() runtime.Object { return &appsapi.PetSetList{} }
+	newListFunc := func() runtime.Object { return &policyapi.PodDisruptionBudgetList{} }
 	storageInterface, dFunc := opts.Decorator(
 		opts.StorageConfig,
-		cachesize.GetWatchCacheSizeByResource(cachesize.PetSet),
-		&appsapi.PetSet{},
+		cachesize.GetWatchCacheSizeByResource(cachesize.PodDisruptionBudget),
+		&policyapi.PodDisruptionBudget{},
 		prefix,
-		petset.Strategy,
+		poddisruptionbudget.Strategy,
 		newListFunc,
 		storage.NoTriggerPublisher,
 	)
 
 	store := &registry.Store{
-		NewFunc: func() runtime.Object { return &appsapi.PetSet{} },
+		NewFunc: func() runtime.Object { return &policyapi.PodDisruptionBudget{} },
 
 		// NewListFunc returns an object capable of storing results of an etcd list.
 		NewListFunc: newListFunc,
-		// Produces a petSet that etcd understands, to the root of the resource
+		// Produces a podDisruptionBudget that etcd understands, to the root of the resource
 		// by combining the namespace in the context with the given prefix
 		KeyRootFunc: func(ctx api.Context) string {
 			return registry.NamespaceKeyRootFunc(ctx, prefix)
 		},
-		// Produces a petSet that etcd understands, to the resource by combining
+		// Produces a podDisruptionBudget that etcd understands, to the resource by combining
 		// the namespace in the context with the given prefix
 		KeyFunc: func(ctx api.Context, name string) (string, error) {
 			return registry.NamespaceKeyFunc(ctx, prefix, name)
 		},
-		// Retrieve the name field of a replication controller
+		// Retrieve the name field of a pod disruption budget
 		ObjectNameFunc: func(obj runtime.Object) (string, error) {
-			return obj.(*appsapi.PetSet).Name, nil
+			return obj.(*policyapi.PodDisruptionBudget).Name, nil
 		},
 		// Used to match objects based on labels/fields for list and watch
-		PredicateFunc:           petset.MatchPetSet,
-		QualifiedResource:       appsapi.Resource("petsets"),
+		PredicateFunc:           poddisruptionbudget.MatchPodDisruptionBudget,
+		QualifiedResource:       policyapi.Resource("poddisruptionbudgets"),
 		EnableGarbageCollection: opts.EnableGarbageCollection,
 		DeleteCollectionWorkers: opts.DeleteCollectionWorkers,
 
 		// Used to validate controller creation
-		CreateStrategy: petset.Strategy,
+		CreateStrategy: poddisruptionbudget.Strategy,
 
 		// Used to validate controller updates
-		UpdateStrategy: petset.Strategy,
-		DeleteStrategy: petset.Strategy,
+		UpdateStrategy: poddisruptionbudget.Strategy,
+		DeleteStrategy: poddisruptionbudget.Strategy,
 
 		Storage:     storageInterface,
 		DestroyFunc: dFunc,
 	}
 	statusStore := *store
-	statusStore.UpdateStrategy = petset.StatusStrategy
+	statusStore.UpdateStrategy = poddisruptionbudget.StatusStrategy
 	return &REST{store}, &StatusREST{store: &statusStore}
 }
 
-// StatusREST implements the REST endpoint for changing the status of an petSet
+// StatusREST implements the REST endpoint for changing the status of an podDisruptionBudget
 type StatusREST struct {
 	store *registry.Store
 }
 
 func (r *StatusREST) New() runtime.Object {
-	return &appsapi.PetSet{}
+	return &policyapi.PodDisruptionBudget{}
 }
 
 // Get retrieves the object from the storage. It is required to support Patch.
