+++ pkg/registry/storage/storageclass/etcd/etcd.go
@@ -22,7 +22,7 @@ import (
 	"k8s.io/kubernetes/pkg/registry/cachesize"
 	"k8s.io/kubernetes/pkg/registry/generic"
 	"k8s.io/kubernetes/pkg/registry/generic/registry"
-	"k8s.io/kubernetes/pkg/registry/storageclass"
+	"k8s.io/kubernetes/pkg/registry/storage/storageclass"
 	"k8s.io/kubernetes/pkg/runtime"
 	"k8s.io/kubernetes/pkg/storage"
 )
